[FEATURE] Add support for PSR-15 HTTP middlewares
(PSR-15) middlewares are hookson steroids. Due to chaining
middlewares and the code request handler they
the can execute code before and after the application,
and may return early, preventing inner middleware and
the application from running at all.

For a proper middleware support we'd the core
and extensions to stop using header()/exit()/die()
or HttpUtility::{redirect,sendRespose}() – they
should all use a PSR-7 Response objects
and ensure those are propagated back all the
way to the Bootstrap, to be sent in Bootstrap::shutdown().

Change-Id: I075639835115f7cf28f18c3814ef6dd190fdf29b
Releases: master
Resolves: #83725