[TASK] EXT:redirect: Reformat to fit common PSR-15 middlewares
No functional change, just rename $currentRequest to $request
(as "current" is not important for a middleware).
Also merge ealry returns (which require each a
`return $handler->handle($request)`) to one,
the final return statement.

This makes it really clear in which situation
the middleware returns an own responsive
or just invokes the next middleware (and returns
their response).

Change-Id: I6851688d6f27c48a4d036ac9898ffe7335317c78
Releases: master
Resolves: #83727