[WIP][FEATURE] Dependency Injection for the core (combined)
This patch is a proof of concept and base for PSR-11 dependency
injection container discussion and therefore marked WIP.

This feature is a combined approach using two containers/concepts:
1) symfony/dependency-injection (exposed as official typo3 api)
   autoconfiguration, autowiring & Service.{yaml,php} config available
2) container-interop/service-provider (provided for internal use)
   (sometimes called PSR-11+)
   factory-style configuration using plain php code

1) Symfony dependency injection is provided to be used by extensions
and throughout the TYPO3 core for autowiring or manually configuration
of services/classes. Extensions may configure symfony's dependency
injection in a file called Configuration/Services.yaml, if they need
more control they may additionally use Configuration/Services.php.
Usage of symfony dependency injection is provided only in non-failsafe
scenarios (currently that means: everything besides the install tool)

2) Service providers may be used (by the core) to provide dependency
injection for services used in the install tool.
container-interop/service-provider is still marked experimental,
therefore the support in TYPO3 is provided for internal use only.

Service providers in failsafe scenario(s) are used through a very
simple, non-compiled container (currently included as dependency: bnf/di).
Symfony's container is too slow when running uncached (as the compilation
needs to be performed on every invocation).
The failsaife container dependency bnf/di could also be bundled into TYPO3
core (if desired).
@todo: we could check whether symfony runtime comilation is fast enough
when not using yaml configuration, reflection, autowiring, and automatic
file lookup. (in other words: configure all required failsafe classes
statically in php files) – like we do with container-inteorp/service-provider.
That effectively means, no difference to this solution, but saving one
dependency.

Services, that are configured through service providers, are
available in the symfony container as well.
They do NOT need to be specified twice. We bridge them to the
symfony container using bnf/service-provider-bridge-bundle.
(Could again be bundled into TYPO3 core, if desired).

Included Examples
-----------------
 * RedirectHandler/RedirectService autowired/autoconfigured though
   symfony di
 * Application classes and their dependencies are handwired,
   This is required as the application runs in falsafe mode when there is
   no configured available (first install)
 * UriBuilder and ConfigurationController are handwired using service
   providers – just as simple demo
 * Middleware configuration is treated as service (like a registry),
   and retrieval is provided through service providers

TODO
----
 * compatibility layer for singleton services for
   GeneralUtility::makeInstance?
   Requires all singleton services to be marked public (in terms of the
   symfony container) :!(
 * default to shared:true? and define shared:false if SingletonInterface?
   Some classes are not explicitly marked as singleton, but they can be
   used as such. Should we add that class to all of them?
 * Provide SymfonyDependencyInjection\Excetion support additionally to
   Configuration/Settings.php? (for extensions?)
   Maybe cleaner than Configuraton/Settings.php, but requires either
   class auto detection (that's what symfony does, it derives the
   extension class name from the bundle name)
 * Extbase container: seperate patch, plan is to replace the extbase
   container (longterm)
   Featurewise symfony missing support for:
   * passing constructor arguments to non singleton classes (protoypes)
   * cyclic dependnecies
 * add container tests in Backend\…\RouteDispatcherTest
 * add tests for AbstractServiceProvider::configureMiddlewares and ::new
 * decide whether AbstractServiceProvider::new() is a good idea
   for handling LoggerAwareInterface (or find alternatives).
   code looks ugly when writing `return static::new($container, $class)`
   well, `GeneralUtlity::makeInstance($class) is not better.
   (maybe drop that, and call setLogger in service provider factories)
 * feature RST files

Depdendency changes
-------------------
composer require \
	container-interop/service-provider:~0.4.0 \
	bnf/di:~0.1.0 \
	symfony/config:^4.1 \
	symfony/dependency-injection:^4.1 \
	bnf/service-provider-bridge-bundle:^3.0

Service Providers – General Information
---------------------------------------
Both containers can be configured through service providers.
Service providers provide factories which perform dependency injection and
extension of services through plain php code. There is no configuration
required besides registering the service provider in composer.json
per extension.
A "service" can be an instantiated class or simple types like arrays or
strings (e.g. to store configuration). Both classes and configuration
values can be (class) dependencies and are thus summarized as "services".

The service provider specification is still expertimental (provided by
container-interop/service-provider). It is supposed to become a PSR.
Therefore all of the support is marked @internal, which allows to adapt
to new releases to this evolving specification, without backward
compatibility problems.

Route and Middleware Dispatcher
-------------------------------
Additionally this patch prepares the middleware dispatcher and the backend
route dispatcher to (optionally) retrieve objects from the container.
These are a good demo for common usecases of the container (and allows to
verify that it is actually working).

Implemented examples:
 * RedirectHandler (frontend middleware)
 * ConfigurationController and InfoModuleController (backend routes)

ServiceProviderInterface defines two methods: getFactories() and
getExtensions(). The former is used to define how classes, or
configuration values are created.
The latter (getExtensions) provides an API to extend services (that means:
services as created by a recipe as returned by getFactrories()).
Use cases for getExtensions():
 * Configuring a signal slot (by extending the signal slot dispatcher)
 * Adding a locallang namespace
 * Adding a configuration setting

Configuration Reading
---------------------
An AbstractServiceProvider is added that implements conventional support
for extending common core services with extension specific settings.
To demo the usecase this patch adds support for configuring middleware
stacks through the container (which processes the service providers).

The concept would be the same for TCA configuration handling:
The TYPO3 core extension provides a factory (e.g. 'middlewares' or 'tca')
which is initially set to be an empty array and all TYPO3 extensions
enhance this through their service provider getExtensions() method.

In case of TCA the TYPO3 convention is to read from Configuration/TCA. We
would implement that convention as a default getExtensions() entry in
AbstractServiceProvider. All ServiceProviders that inherit from that
abstract class thus provide support for the configuration convention.
We add a LegacyServiceProvider which is used, in case a TYPO3 extension
does not define an own service provider. This is to ensure those
extensions still benefit from the default getExtensions() method(s)
in AbstractServiceProvider.

The logic to initialize configurations like $GLOBALS['TYPO3_CONF_VARS']
and $GLOBALS['TCA'] should be moved into the AbstractServiceProvider in
far future (as that will be breaking).

Releases: master
Resolves: #84112
Change-Id: I8efd817066b528a5953c56fdd027cb94b2bb8eb3