[TASK] Merge the package DependencyResolver into the PackageManager
The DependencyResolver and the PackageManager have a cyclic dependency to
each other which is currently resolved using GeneralUtility::makeInstance.
As the DependencyResolver is actually only used for the
PackageManager – and relies on it – it can be merged, saving a lot of
hassle.

The DepenencyResolver class in not marked @internal and is therefore
deprecated.

Releases: master
Resolves: #84109
Change-Id: I71adccec3f13eb6de859f065937fbcde369758fe