[TASK] Workaround a cyclic dependency in QueryFactory <-> DataMapper
This can not be resolved by symfony currently (or I am doing
something wrong). Delay the cyclic dependency a bit.
We either need to get symfony to support cyclic dependnecies
like these, or the migration to symfony will be a breaking change (for
extensions that use cyclic dependencies using injections)

Change-Id: Ief0b3c9455f86a7bccc88c755023e6359fb5d5f9