[WIP][TASK] Extbase: use serviceLocator in various locations (start with the Dispatcher)
We may use these instead of ObjectManager->get(), as the can
only return those services, that are intended for the usecase.
(Provides an explicit dependency graph, instead of implicitly
depending on ALL public services).

using methods described in
https://symfony.com/doc/current/service_container/service_subscribers_locators.html#defining-a-service-locator
it should be possible to define a serviceLocator
that allows to pass only a subset of services to the extbase Dispatcher (only controllers)

Change-Id: Ic780e29cf71bbec1d7582a0196a846fb094c801e