[WIP][TASK] Inject singletons used by EXT:install in service providers
This will allow both, dependency injection for these services (manually
wired in service providers), and usage in install/maintenance tool
(where we do not use the caching symfony container for basic tasks).

This move to DI is possible thanks to the failsafe container and the
service providers which can fed the failsafe container with service
factories, where we can wire services manually.

Note: The definition in a service provider means we are required to use
manual wiring and are forced to define the dependencies in the service
provider when we add new depenencies to services that are being used
by the install..
With that approach we can assure that we do not accidentally add
dependencies to services which are used by the install tool without
defining a factory in a service provider.

TODO:
 * Use dependency injection in all install tool services/controllers.
   The patch will grow, but it allows to understand, why we need to
   add certain core classes to service providers
 * Test all cases/methods in install tool
 * Remove debug LogicException in GeneralUtility::makeInstance

Notes:

a) GLOBALS['LANG'] mocks in
EXT:core/Tests/Unit/DataHandling/DataHandlerTest.php have been removed
as the code under test does no longer use
$GLOBALS['LANG']->csConvObj->substr() but mb_substr.
The test had been introduced in
https://review.typo3.org/c/Packages/TYPO3.CMS/+/42260/
but the tested code was adapted in
https://review.typo3.org/50607

b) We got a bit ugly constructs now, where a unit(!) test used
    $GLOBALS['LANG'] = new \TYPO3\CMS\Core\Localization\LanguageService();
we now have to encode the dependency structure:
    $GLOBALS['LANG'] = new LanguageService(new Locales,
        new LocalizationFactory(
            new LanguageStore,
            $cacheManagerProphecy->reveal()
        )
    );
This isn't nice, but this change reveals that the unit(!) tests should
either be adapted, removed or be moved to a functional test.
Such adaption are out of scope for this change.

Releases: master
Resolves: #89892
Resolves: #89891
Change-Id: Ib72d6440f81b2c0d05279e8768697c3b48aecfe4