[BUGFIX] Resolve and link modified slugs in workspaces properly
When a page (slug) is modified in a preview workspace,
links to that page need to be changed to the new slug, as the
preview (PageRouter) should evaluate this as well.

1) The PageRouter should include records with "-1"
to find them as well, but then fix the "pid" again to
find the pid of the online version.

2) The PageLink Builder now uses the SiteFinder nstead of SiteMatcher
to detect whether the linked page uid is part of a configured site.
We do not need to create/detect a PseudoSite here, only instances
of Site are of interest.

Resolves: #87871
Releases: master, 9.5
Change-Id: Ifd6add71bec1616049f8c6a50a42bc9f573395e2