[TASK] Add be_user id to LinkGenerator page-version test context
Some of the previous todos were failing because the test context
contained a workspace id, but missed a backend user id.
In that case the workspace id was ignored ignored by the test
framework (see TYPO3\JsonResponse\Middleware\BackendUserMiddleware).

This patch also adapts the SlugLinkGeneratorTest to check link
generation to versioned record both with and without a backend user
authentication.

Releases: master, 9.5
Resolves: #87973
Change-Id: I0a2ec41234963fd8af0ed3adf951f139f1e85215