[TASK] Hook up ContainerInterface in backend route dispatcher
Transform InfoModuleController and it's dependency ModuleTemplate
into a symfony manageed services  to retrieve dependencies from
symfony instead of GeneralUtility::makeInstance.
ModuleTemplate is a prototype and is therefore marked shared: false.
It is marked public: true so legacy calls to
GeneralUtility::makeInstance(ModuleTemplate::class) resort to the
container and properly inject dependencies.

Also add a backend.controller symfony tag, to be universally
applied for backend controllers. This tag automatically configures
the controller to be publicly available from $container->get()
which allows the route dispatcher to lazily instantiate
the controller.

Releases: master
Resolves: #88721
Change-Id: I076c306736243e693542f2774dbe1108e28fe731