[FEATURE] Provide implementation for PSR-17 HTTP Message Factories
Support for PSR-17 HTTP Message Factories has been added.

PSR-17 HTTP Factories are intended to be used by PSR-15 request handlers
in order to create a PSR-7 compatible message objects.

Classes may use dependency injection to use any of the available PSR-17
HTTP Factory interfaces.

The Request/Response base class Message is adapted to be able to lazily
initialize a stream when getBody() is called.
This is done as the (Stream)RequestFactoryInterface does not allow
to control Stream properties. Therefore it's a performance
optimization to defer initialization as it is likely, that a new
Stream with be added with withStream() anyway. (Which would mean
resources for the intermediate stream would have been wasted)

Furthermore some DocBlocks are adapted to reflect the variadic
UriInterface/StreamInterface parameters that are already handled in
code but were not documented. These cases are actually needed/required
by the PSR-17 factory implementation now.

composer require psr/http-factory:^1.0

Releases: master
Resolves: #89018
Change-Id: Ie6b9d865679bbf6f5d3d030b0ed1a3f277c47a3d