[TASK] Rework TypeScript to be 3.6 compatible
After shedding tears, blood and sacrificing brain cells and CPU time, we
finally managed to upgrade the code base to a recent TypeScript version.

Interestingly, the current used version 2.4 is not that strict about
typings, allowing potential bugs in the code.

With this patch, our code can be compiled with TypeScript 3.6.
The dependency will be raised with another commit.

Resolves: #89104
Releases: master
Change-Id: I023ba89b4d4d8d98394d046c53db7fa64c12d44c