[BUGFIX] Fix compatibility with symfony/dependency-injection:5.0
The recently released symfony 4.4 introduced a deprecation for
non-object services and so 5.0 dropped support.
That means we run into exception with symfony 5.0 when
we add synthetic non-object services. Symfony 5.0 contains
a type declaration that allows for objects only.

We provided 4 boolean services in v10.0 but never used
or documented these in core:
 - env.is_unix
 - env.is_windows
 - env.is_cli
 - env.is_compoer_mode

They can actually be substitued by using Symfony env pass
(that is supported due to our EnvVarProcessor):
 - "%env(TYPO3:isUnix)%"
 - "%env(TYPO3:isWindows)%"
 - "%env(TYPO3:isCli)%"
 - "%env(TYPO3:isComposerMode)%"

One more was boolean service was actually used in core
and is now migrated into the internal boot.state service.
 - cache.disabled

Releases: master
Resolves: #
Change-Id: If40ef73fdce61a81732b9670876aec7c457648d5