[FEATURE] Dashboard for TYPO3
A dashboard is introduced into TYPO3 to show the most important
information to the current logged in user.

Every user with access to this backend module can now have one
or more personal dashboards. Each dashboard can contain several
widgets. Which widgets and in which order the widgets are shown
is up to the users themselves.

Resolves: #90333
Releases: master
Change-Id: I964a52846ec9a1c901baa89508b9df0caecdf81f