[WIP][TASK] Add RequestStack
Inspired by symfony.
Nothing that is suggested to be used whenever avoidable.

Automatically enhances the TYPO3_REQUEST compat layer when middlewares
dispatch a new/changes PSR-7 request.

 * TODO adapt tests
 * Think and write about the rationale

This is mostly an experiment to see how a request stack would work n TYPO3.
There is not intent to actually implement this yet.

Change-Id: I09e98d1b6c881d398d5bdd79cfe5e03463cb10a7
Releases: master
Resolves: #