[!!!][TASK] Remove deprecated constructor arguments in TSFE
Since TYPO3 v10 the god-object TSFE is now built with
everything that is mandatory for rendering the Frontend:
* Context
* Site
* SiteLanguage
* PageArguments
* FrontendUser

The ultimate goal is to make TSFE be built without any global logic
in the future, so it can be re-used and simulated in the backend as well,
if needed. Also, the goal is to reduce the access to the global object,
relying on other factors instead.

Resolves: #91511
Related: #91473
Releases: master
Change-Id: Ic237b43be1e98e18477f2de1d63cb80a1c2ab241