[FEATURE] Support autowired LoggerInterface injection
Logger instances may be required to be available during object
construction, LoggerAwareInterface isn't an option in that case.
Therefore LoggerInterface as constructor argument is now
autowired (if the service is configured to use autowiring)
and instantiated with an object-specific logger.

We add a symfony compiler pass that prepares an inline factory which
fetches the appropriate Logger instance from the LogManager.
One of the TYPO3 core classes that require this is the Scheduler.

Scheduler used to contain a safety net which fetches a logger instance
from LogManager in case the logger is already needed during the
constructor invocation. In that case the logger was not available, as it
was injected via setter injection (setLogger) later
(per LoggerAwareInterface).

Releases: master
Resolves: #95044
Change-Id: I57539a731a305da6abc092d2383e2d2f4c1d985c